move testdata
