+++ pkg/webhook/operatingsystemconfig/testdata/00-base-config.rules
