+++ pkg/webhook/operatingsystemconfig/testdata/10-privilege-escalation.rules
