+++ pkg/webhook/operatingsystemconfig/testdata/11-privileged-special.rules
