+++ pkg/webhook/operatingsystemconfig/testdata/12-system-integrity.rules
