+++ pkg/webhook/operatingsystemconfig/testdata/60-audit-with-tls.conf
