+++ pkg/webhook/operatingsystemconfig/testdata/60-audit.conf
