+++ pkg/webhook/operatingsystemconfig/testdata/audit/audit-configmap.yaml
