+++ pkg/webhook/operatingsystemconfig/testdata/configure-rsyslog.sh
