+++ pkg/webhook/operatingsystemconfig/testdata/process-rsyslog-pstats.sh
