+++ pkg/webhook/operatingsystemconfig/testdata/rsyslog-config-simple.conf.tpl
